Renamed NutrientCircles.js file.
